backend/pyproject.toml
[project]
name = "factible"
version = "0.1.0"
description = "E2E project to fact check social media posts"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-ai[examples]>=1.0.0,<2.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "youtube-transcript-api>=1.2.2,<2.0.0",
    "requests>=2.25.0,<3.0.0",
    "selenium>=4.15.2,<5.0.0",
    "python-whois>=0.8.0,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "typer>=0.15.0,<1.0.0",
    "httpx>=0.28.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
]

[project.scripts]
factible-experiments = "experiments.cli:app"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0,<5.0.0",
    "types-requests>=2.32.4",
    "types-pyyaml>=6.0.0",
    "py-spy>=0.4.1,<0.5.0",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "pandas>=2.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.12"
show_error_codes = true
install_types = true
non_interactive = true
